pyproject.toml
[tool.poetry]
name = "fast-api-assignment"
version = "0.1.0"
description = ""
authors = ["Saeem Mehmood <saimmahmood61@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.112.1"
uvicorn = "^0.30.6"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pymongo = "^4.8.0"
python-dotenv = "^1.0.1"
python-jose = {extras = ["crypto"], version = "^3.3.0"}
python-multipart = "^0.0.9"
pytest = "^8.3.2"
httpx = "^0.27.0"
pandas = "^2.2.2"
openpyxl = "^3.1.5"
sentry-sdk = {extras = ["fastapi"], version = "^2.13.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
